// src/app/models/lacre/lacre.ts
export class Lacre {
    auto: string;
    data: Date;
    grupo: string;
    quantidade: string;
    lacre: string;
    linha: number;
    numero: string;
    pos: number;
    processo: string;
    trm: string;
    codigo: string;
    recebedor: string;
    status: string;
    acao: string;
}